Initial commit for HTMark
This is a fun/quick little script that will save URIs in HTML files. I
got the idea from HTMTied[1].

Basically, what it does is stick the URI in a HTML file that, when
opened, will redirect to that URI. By saving your "bookmarks" in HTML
files they'll be easily hand-editable, trackable by git, portable to
any OS, and manageable by a file manager of your choice.

I kept the implementation pretty simple and avoided any extra
features/dependencies on purpose; if we want fancier management we will
need to make a separate project that will pull HTMark in.

All this script will really do is see if the URI is valid and
interpolate it into a template. Both of these things are easily done in
the MRI's standard library.

Enjoy!

[1] http://www.aurelitec.com/htmtied/windows/